GetHtml2 is a modified version of GetHtml.
GetHtml2 is given 5 attempts to extract the html from the page.

Previously, rate limiting was occurring, so for every attempt a new chromedp context is created, the attempt will check for errors (empty html or an error page), sleeping when necessary before retrying and returning the html content.

The sleep period appears to be sufficient enough to avoid continuous rate limiting.